+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Master Dev Branch
 
+### 2025-10-21
+- Added top reports by Volume
+
 ## Version v1.6.7 - Released: 2025-10-20
 
 ### 2025-10-20